+++ bindings/nostr-sdk-js/src/nwc/mod.rs
@@ -17,6 +17,7 @@ use crate::protocol::nips::nip47::{
     JsLookupInvoiceResponseResult, JsMakeInvoiceRequestParams, JsMakeInvoiceResponseResult,
     JsNostrWalletConnectURI, JsPayKeysendRequestParams, JsPayKeysendResponseResult,
 };
+use crate::relay::JsRelayStatus;
 
 #[wasm_bindgen]
 extern "C" {
@@ -55,6 +56,11 @@ impl JsNwc {
         Self {
             inner: NWC::with_opts(uri.deref().clone(), opts.deref().clone()),
         }
+    }
+
+    /// Get relay status
+    pub fn status(&self) -> JsRelayStatus {
+        self.inner.status().into()
     }
 
     /// Pay invoice